+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Added a `new_blog_post()` function (thanks, @topeto, #22). 
 
+- Make `quarto_render(as_job = TRUE)` wrapable (thanks, @salim-b, #105).
+
 - Quarto CLI will now correctly use the same R version than the one used to run functions in this package (#204).
 
 - Add `quarto_available()` function to check if Quarto CLI is found (thanks, @hadley, #187).